Split reduction routines from math.cc
