+++ chainerx_cc/chainerx/cuda/cuda_device/dot.cu
@@ -24,8 +24,8 @@
 #include "chainerx/float16.h"
 #include "chainerx/kernels/creation.h"
 #include "chainerx/kernels/linalg.h"
-#include "chainerx/kernels/math.h"
 #include "chainerx/kernels/misc.h"
+#include "chainerx/kernels/reduction.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
