+++ chainerx_cc/chainerx/kernels/math.h
@@ -3,7 +3,6 @@
 #include <cstdint>
 
 #include "chainerx/array.h"
-#include "chainerx/axes.h"
 #include "chainerx/kernel.h"
 #include "chainerx/scalar.h"
 
@@ -196,20 +195,6 @@ public:
     static const char* name() { return "IsFinite"; }
 
     virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-// Calculate the sum of an array.
-// It will be summed over the specified axes.
-// `axis` must be normalized so that
-// - it has only positive values,
-// - it is sorted, and
-// - it has no duplicated values.
-// Otherwise, the behavior is undefined.
-class SumKernel : public Kernel {
-public:
-    static const char* name() { return "Sum"; }
-
-    virtual void Call(const Array& a, const Axes& axis, const Array& out) = 0;
 };
 
 // Compares x1 and x2 and assign either pos or neg according to the result.
