+++ chainerx_cc/chainerx/python/routines.cc
@@ -28,6 +28,7 @@
 #include "chainerx/routines/math.h"
 #include "chainerx/routines/normalization.h"
 #include "chainerx/routines/pooling.h"
+#include "chainerx/routines/reduction.h"
 #include "chainerx/routines/sorting.h"
 #include "chainerx/routines/statistics.h"
 #include "chainerx/routines/trigonometric.h"
