+++ chainerx_cc/chainerx/routines/connection.cc
@@ -24,7 +24,7 @@
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/linalg.h"
-#include "chainerx/routines/math.h"
+#include "chainerx/routines/reduction.h"
 #include "chainerx/routines/type_util.h"
 #include "chainerx/stack_vector.h"
