+++ chainerx_cc/chainerx/routines/math.h
@@ -5,7 +5,6 @@
 #include <nonstd/optional.hpp>
 
 #include "chainerx/array.h"
-#include "chainerx/axes.h"
 #include "chainerx/scalar.h"
 
 namespace chainerx {
@@ -72,8 +71,6 @@ Array TrueDivide(Scalar x1, const Array& x2);
 
 Array Reciprocal(const Array& x);
 
-Array Sum(const Array& a, const OptionalAxes& axis = nonstd::nullopt, bool keepdims = false);
-
 Array Maximum(const Array& x1, Scalar x2);
 Array Maximum(Scalar x1, const Array& x2);
 Array Maximum(const Array& x1, const Array& x2);
@@ -86,19 +83,9 @@ Array Exp(const Array& x);
 Array Log(const Array& x);
 Array Log10(const Array& x);
 
-// Returns the LogSumExp (LSE) of x, reduced along the specified axes.
-// If no axes are specified, all axes will be reduced.
-Array LogSumExp(const Array& x, const OptionalAxes& axis = nonstd::nullopt, bool keepdims = false);
-
-// Returns the logarithm of the softmax of x along the specified axes.
-// If no axes are specified, the softmax is applied on the second axis.
-Array LogSoftmax(const Array& x, const OptionalAxes& axis = nonstd::nullopt);
-
 Array Sigmoid(const Array& x);
 
 Array Relu(const Array& x);
-
-Array Softmax(const Array& x, const OptionalAxes& axis = nonstd::nullopt);
 
 Array Square(const Array& x);
