+++ chainerx_cc/examples/mnist/train_mnist.cc
@@ -21,6 +21,7 @@
 #include "chainerx/routines/linalg.h"
 #include "chainerx/routines/manipulation.h"
 #include "chainerx/routines/math.h"
+#include "chainerx/routines/reduction.h"
 #include "chainerx/shape.h"
 #include "chainerx/slice.h"
 #include "mnist.h"
